Convert `mainexample.jl` to documentation, add tests, and reorganize the source files
