+++ .gitignore
@@ -1,2 +1,8 @@
-*DS_Store
+*.jl.*.cov
+*.jl.cov
+*.jl.mem
+
+/docs/build/
+
+.DS_Store
 Manifest.toml
