+++ LICENSE
@@ -1,5 +1,4 @@
-Copyright (c) 2019 Thibaut Lienart
-Copyright (c) 2020 Chad Scherrer
+Copyright (c) 2020 Chad Scherrer, Thibaut Lienart, Dilum Aluthge, and contributors
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
