+++ Project.toml
@@ -1,6 +1,6 @@
 name = "SossMLJ"
 uuid = "cf2dc0df-ab57-42da-b773-28c60db69bd5"
-authors = ["Chad Scherrer"]
+authors = ["Chad Scherrer", "Thibaut Lienart", "Dilum Aluthge", "contributors"]
 version = "0.1.0"
 
 [deps]
@@ -9,21 +9,23 @@ MLJBase = "a7f614a8-145f-11e9-1d2a-a57a1082229d"
 MLJModelInterface = "e80e1ace-859a-464e-9ed9-23947d8ae3ea"
 MonteCarloMeasurements = "0987c9cc-fe09-11e8-30f0-b96dd679fdca"
 Soss = "8ce77f84-9b61-11e8-39ff-d17a774bf41c"
-StatsFuns = "4c63d2b9-4356-54db-8cca-17b64c39e42c"
 Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
 
 [compat]
 Distributions = "0.23"
+Literate = "2"
 MLJ = "0.12"
-MLJModelInterface = "0.3"
+MLJBase = "0.14"
+MLJModelInterface = "0.3.3"
 MonteCarloMeasurements = "0.9"
 Soss = "0.13"
 Tables = "1"
-julia = "1.3"
+julia = "1.4"
 
 [extras]
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
+Literate = "98b081ad-f1c9-55d3-8b20-4c87d4299306"
 MLJ = "add582a8-e3ab-11e8-2d5e-e98b27df1bc7"
 MLJModelInterface = "e80e1ace-859a-464e-9ed9-23947d8ae3ea"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
@@ -31,4 +33,4 @@ Soss = "8ce77f84-9b61-11e8-39ff-d17a774bf41c"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["Distributions", "Documenter", "MLJ", "MLJModelInterface", "Random", "Soss", "Test"]
+test = ["Distributions", "Documenter", "Literate", "MLJ", "MLJModelInterface", "Random", "Soss", "Test"]
