+++ docs/src/api.md
@@ -1,0 +1,8 @@
+# API
+
+```@index
+```
+
+```@autodocs
+Modules = [SossMLJ]
+```
