+++ docs/src/index.md
@@ -4,10 +4,3 @@
 [Soss](https://github.com/cscherrer/Soss.jl) probabilistic programming library
 into the [MLJ](https://github.com/alan-turing-institute/MLJ.jl) machine
 learning framework.
-
-```@index
-```
-
-```@autodocs
-Modules = [SossMLJ]
-```
